compiler: Refactor __padding_setup__
